moved Affilation to faction project
