+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/entities/GameWorld.kt
@@ -1,7 +1,7 @@
 package de.gleex.pltcmd.game.ui.entities
 
 import de.gleex.pltcmd.game.engine.entities.types.*
-import de.gleex.pltcmd.model.elements.Affiliation
+import de.gleex.pltcmd.model.faction.Affiliation
 import de.gleex.pltcmd.model.world.Sector
 import de.gleex.pltcmd.model.world.WorldMap
 import de.gleex.pltcmd.model.world.coordinate.Coordinate
