+++ game/ui/src/main/kotlin/de/gleex/pltcmd/game/ui/entities/TileRepository.kt
@@ -2,8 +2,8 @@ package de.gleex.pltcmd.game.ui.entities
 
 import de.gleex.pltcmd.game.ui.strings.Format
 import de.gleex.pltcmd.game.ui.strings.extensions.toFrontendString
-import de.gleex.pltcmd.model.elements.Affiliation
 import de.gleex.pltcmd.model.elements.CommandingElement
+import de.gleex.pltcmd.model.faction.Affiliation
 import de.gleex.pltcmd.model.signals.core.SignalStrength
 import de.gleex.pltcmd.model.world.terrain.Terrain
 import de.gleex.pltcmd.model.world.terrain.TerrainHeight
