+++ model/elements/pom.xml
@@ -14,6 +14,10 @@
 
     <dependencies>
         <dependency>
+            <groupId>de.gleex.pltcmd.model</groupId>
+            <artifactId>faction</artifactId>
+        </dependency>
+        <dependency>
             <groupId>de.gleex.pltcmd.util</groupId>
             <artifactId>namegeneration</artifactId>
         </dependency>
