+++ model/faction/src/main/kotlin/de/gleex/pltcmd/model/faction/Affiliation.kt
@@ -1,4 +1,4 @@
-package de.gleex.pltcmd.model.elements
+package de.gleex.pltcmd.model.faction
 
 /**
  * The affiliation of an element describes how a faction sees it. This may differ for the same element seen by different factions.
