+++ model/pom.xml
@@ -39,6 +39,11 @@
             </dependency>
             <dependency>
                 <groupId>de.gleex.pltcmd.model</groupId>
+                <artifactId>faction</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>de.gleex.pltcmd.model</groupId>
                 <artifactId>mapgeneration</artifactId>
                 <version>${project.version}</version>
             </dependency>
